keepalived: add option to override service running check
some init.d scripts like firewall and sqm do not return the actual state of the service if called with "running" parameter. This result in the init script called with "start" parameter and the service may not load the new configuration. firewall init script is one of this
An option is added in order to skip the "running" check for the service

Signed-off-by: Francesco Benini <francy.benini@gmail.com>