keepalived: fix ambigiuos set_reload_if_sync and set_update_target
when "set_reload_if_sync" is not set in the hotplug script, the service is not expected to reload. That is not true because even if not set, the value is set to the default 1 (reload active) or equals the parameter set when "keepalived_hotplug" is called.
The default behavior should be:
    - reload if set_reload_if_sync is called
    - NOT reload if set_reload_if_sync is NOT called
A similar fix is ported to "set_update_target"

Signed-off-by: Francesco Benini <francy.benini@gmail.com>